+++ src/domains/TotalLotto.js
@@ -1,6 +1,6 @@
 import CONSTANT from '../constants/costant.js';
 import { MissionUtils } from '@woowacourse/mission-utils';
-import Lotto from '../Lotto.js';
+import Lotto from './Lotto.js';
 
 class TotalLotto {
   #lottos;
